services/backend/game3d/tools/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.5
daphne==4.1.0
Django==5.0.4
django-rest-swagger==2.2.0
djangorestframework==3.15.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
msgpack==1.0.8
openapi-codec==1.3.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
redis==5.0.3
requests==2.31.0
service-identity==24.1.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
zope.interface==6.2
